# Use C++20 and swl::variant instead of std::variant.