--- a/.github/actions/install-unreal-macos/exclude-from-unzip.lst
+++ b/.github/actions/install-unreal-macos/exclude-from-unzip.lst
@@ -8,7 +8,6 @@ UE_5.2/Engine/Binaries/Mac/UnrealGame
 UE_5.2/Engine/Binaries/Mac/UnrealGame-Mac-Shipping
 UE_5.2/Engine/Binaries/Mac/*.dSYM
 UE_5.2/Engine/Plugins/Experimental/NNERuntimeIREE/*
-UE_5.2/Engine/Plugins/Experimental/TextToSpeech/*
 UE_5.2/Engine/Plugins/Experimental/Avalanche/*
 UE_5.2/Engine/Plugins/Runtime/NetworkPredictionExtras/*
 UE_5.2/Engine/Plugins/Runtime/Metasound/*
@@ -23,7 +22,6 @@ UE_5.3/Engine/Binaries/Mac/UnrealGame
 UE_5.3/Engine/Binaries/Mac/UnrealGame-Mac-Shipping
 UE_5.3/Engine/Binaries/Mac/*.dSYM
 UE_5.3/Engine/Plugins/Experimental/NNERuntimeIREE/*
-UE_5.3/Engine/Plugins/Experimental/TextToSpeech/*
 UE_5.3/Engine/Plugins/Experimental/Avalanche/*
 UE_5.3/Engine/Plugins/Runtime/NetworkPredictionExtras/*
 UE_5.3/Engine/Plugins/Runtime/Metasound/*
@@ -40,7 +38,6 @@ UE_5.4/Engine/Binaries/Mac/UnrealGame-Mac-DebugGame
 UE_5.4/Engine/Binaries/Mac/UnrealGame
 UE_5.4/Engine/Binaries/Mac/UnrealGame-Mac-Shipping
 UE_5.4/Engine/Plugins/Experimental/NNERuntimeIREE/*
-UE_5.4/Engine/Plugins/Experimental/TextToSpeech/*
 UE_5.4/Engine/Plugins/Experimental/Avalanche/*
 UE_5.4/Engine/Plugins/Experimental/MoverExamples/*
 UE_5.4/Engine/Plugins/Experimental/MoverTests/*