--- a/.github/actions/install-unreal-macos/exclude-from-unzip.lst
+++ b/.github/actions/install-unreal-macos/exclude-from-unzip.lst
@@ -42,14 +42,6 @@ UE_5.4/Engine/Plugins/Experimental/Avalanche/*
 UE_5.4/Engine/Plugins/Experimental/MoverExamples/*
 UE_5.4/Engine/Plugins/Experimental/MoverTests/*
 UE_5.4/Engine/Plugins/Experimental/Mutable/*
-UE_5.4/Engine/Plugins/Runtime/NetworkPredictionExtras/*
-UE_5.4/Engine/Plugins/Runtime/Metasound/*
-UE_5.4/Engine/Plugins/Runtime/GooglePAD/*
-UE_5.4/Engine/Plugins/Runtime/Harmonix/*
-UE_5.4/Engine/Plugins/Importers/USDImporter/*
-UE_5.4/Engine/Plugins/Animation/*
-UE_5.4/Engine/Plugins/NNE/*
-UE_5.4/Engine/Plugins/Media/AjaMedia/*
 UE_5.4/Engine/Source/Runtime/OodleDataCompression/*
 UE_5.4/Engine/Binaries/ThirdParty/DotNet/6.0.302/mac-x64/*
 UE_5.4/Engine/Binaries/ThirdParty/CEF3/*