--- a/.github/actions/install-unreal-macos/exclude-from-unzip.lst
+++ b/.github/actions/install-unreal-macos/exclude-from-unzip.lst
@@ -38,7 +38,6 @@ UE_5.4/Engine/Plugins/Experimental/MoverTests/*
 UE_5.4/Engine/Plugins/Experimental/Mutable/*
 UE_5.4/Engine/Binaries/ThirdParty/DotNet/6.0.302/mac-x64/*
 UE_5.4/Engine/Binaries/ThirdParty/CEF3/*
-UE_5.4/Engine/Binaries/ThirdParty/Python3/*
 UE_5.4/Engine/Binaries/DotNET/AutomationTool/AutomationScripts/Platforms/Windows/*
 UE_5.4/Engine/Intermediate/ScriptModules/Win.Automation.json
 UE_5.4/Engine/Binaries/DotNET/AutomationTool/AutomationScripts/Platforms/Android/*