--- a/.github/actions/install-unreal-macos/exclude-from-unzip.lst
+++ b/.github/actions/install-unreal-macos/exclude-from-unzip.lst
@@ -31,11 +31,6 @@ UE_5.4/Engine/Binaries/Mac/UnrealEditor-Mac-DebugGame.app/*
 UE_5.4/Engine/Binaries/Mac/UnrealGame-Mac-DebugGame
 UE_5.4/Engine/Binaries/Mac/UnrealGame
 UE_5.4/Engine/Binaries/Mac/UnrealGame-Mac-Shipping
-UE_5.4/Engine/Plugins/Experimental/NNERuntimeIREE/*
-UE_5.4/Engine/Plugins/Experimental/Avalanche/*
-UE_5.4/Engine/Plugins/Experimental/MoverExamples/*
-UE_5.4/Engine/Plugins/Experimental/MoverTests/*
-UE_5.4/Engine/Plugins/Experimental/Mutable/*
 UE_5.4/Engine/Binaries/ThirdParty/DotNet/6.0.302/mac-x64/*
 UE_5.4/Engine/Binaries/ThirdParty/CEF3/*
 UE_5.4/Engine/Binaries/DotNET/AutomationTool/AutomationScripts/Platforms/Windows/*