--- a/Source/CesiumRuntime/Private/CesiumMetadataValue.cpp
+++ b/Source/CesiumRuntime/Private/CesiumMetadataValue.cpp
@@ -51,7 +51,7 @@ UCesiumMetadataValueBlueprintLibrary::GetTrueComponentType(
 bool UCesiumMetadataValueBlueprintLibrary::GetBoolean(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     bool DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [DefaultValue](auto value) -> bool {
         return CesiumGltf::MetadataConversions<bool, decltype(value)>::convert(
                    value)
@@ -65,7 +65,7 @@ PRAGMA_ENABLE_DEPRECATION_WARNINGS
 uint8 UCesiumMetadataValueBlueprintLibrary::GetByte(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     uint8 DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [DefaultValue](auto value) -> uint8 {
         return CesiumGltf::MetadataConversions<uint8, decltype(value)>::convert(
                    value)
@@ -77,7 +77,7 @@ uint8 UCesiumMetadataValueBlueprintLibrary::GetByte(
 int32 UCesiumMetadataValueBlueprintLibrary::GetInteger(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     int32 DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [DefaultValue](auto value) {
         return CesiumGltf::MetadataConversions<int32, decltype(value)>::convert(
                    value)
@@ -89,7 +89,7 @@ int32 UCesiumMetadataValueBlueprintLibrary::GetInteger(
 int64 UCesiumMetadataValueBlueprintLibrary::GetInteger64(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     int64 DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [DefaultValue](auto value) -> int64 {
         return CesiumGltf::MetadataConversions<int64, decltype(value)>::convert(
                    value)
@@ -101,7 +101,7 @@ int64 UCesiumMetadataValueBlueprintLibrary::GetInteger64(
 float UCesiumMetadataValueBlueprintLibrary::GetFloat(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     float DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [DefaultValue](auto value) -> float {
         return CesiumGltf::MetadataConversions<float, decltype(value)>::convert(
                    value)
@@ -113,7 +113,7 @@ float UCesiumMetadataValueBlueprintLibrary::GetFloat(
 double UCesiumMetadataValueBlueprintLibrary::GetFloat64(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     double DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [DefaultValue](auto value) -> double {
         return CesiumGltf::MetadataConversions<double, decltype(value)>::
             convert(value)
@@ -125,7 +125,7 @@ double UCesiumMetadataValueBlueprintLibrary::GetFloat64(
 FIntPoint UCesiumMetadataValueBlueprintLibrary::GetIntPoint(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     const FIntPoint& DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [&DefaultValue](auto value) -> FIntPoint {
         if constexpr (CesiumGltf::IsMetadataString<decltype(value)>::value) {
           return UnrealMetadataConversions::toIntPoint(value, DefaultValue);
@@ -142,7 +142,7 @@ FIntPoint UCesiumMetadataValueBlueprintLibrary::GetIntPoint(
 FVector2D UCesiumMetadataValueBlueprintLibrary::GetVector2D(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     const FVector2D& DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [&DefaultValue](auto value) -> FVector2D {
         if constexpr (CesiumGltf::IsMetadataString<decltype(value)>::value) {
           return UnrealMetadataConversions::toVector2D(value, DefaultValue);
@@ -159,7 +159,7 @@ FVector2D UCesiumMetadataValueBlueprintLibrary::GetVector2D(
 FIntVector UCesiumMetadataValueBlueprintLibrary::GetIntVector(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     const FIntVector& DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [&DefaultValue](auto value) -> FIntVector {
         if constexpr (CesiumGltf::IsMetadataString<decltype(value)>::value) {
           return UnrealMetadataConversions::toIntVector(value, DefaultValue);
@@ -176,7 +176,7 @@ FIntVector UCesiumMetadataValueBlueprintLibrary::GetIntVector(
 FVector3f UCesiumMetadataValueBlueprintLibrary::GetVector3f(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     const FVector3f& DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [&DefaultValue](auto value) -> FVector3f {
         if constexpr (CesiumGltf::IsMetadataString<decltype(value)>::value) {
           return UnrealMetadataConversions::toVector3f(value, DefaultValue);
@@ -193,7 +193,7 @@ FVector3f UCesiumMetadataValueBlueprintLibrary::GetVector3f(
 FVector UCesiumMetadataValueBlueprintLibrary::GetVector(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     const FVector& DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [&DefaultValue](auto value) -> FVector {
         if constexpr (CesiumGltf::IsMetadataString<decltype(value)>::value) {
           return UnrealMetadataConversions::toVector(value, DefaultValue);
@@ -210,7 +210,7 @@ FVector UCesiumMetadataValueBlueprintLibrary::GetVector(
 FVector4 UCesiumMetadataValueBlueprintLibrary::GetVector4(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     const FVector4& DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [&DefaultValue](auto value) -> FVector4 {
         if constexpr (CesiumGltf::IsMetadataString<decltype(value)>::value) {
           return UnrealMetadataConversions::toVector4(value, DefaultValue);
@@ -227,7 +227,7 @@ FVector4 UCesiumMetadataValueBlueprintLibrary::GetVector4(
 FMatrix UCesiumMetadataValueBlueprintLibrary::GetMatrix(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     const FMatrix& DefaultValue) {
-  auto maybeMat4 = std::visit(
+  auto maybeMat4 = swl::visit(
       [&DefaultValue](auto value) -> std::optional<glm::dmat4> {
         return CesiumGltf::MetadataConversions<glm::dmat4, decltype(value)>::
             convert(value);
@@ -241,7 +241,7 @@ FMatrix UCesiumMetadataValueBlueprintLibrary::GetMatrix(
 FString UCesiumMetadataValueBlueprintLibrary::GetString(
     UPARAM(ref) const FCesiumMetadataValue& Value,
     const FString& DefaultValue) {
-  return std::visit(
+  return swl::visit(
       [&DefaultValue](auto value) -> FString {
         using ValueType = decltype(value);
         if constexpr (
@@ -262,7 +262,7 @@ FString UCesiumMetadataValueBlueprintLibrary::GetString(
 
 FCesiumPropertyArray UCesiumMetadataValueBlueprintLibrary::GetArray(
     UPARAM(ref) const FCesiumMetadataValue& Value) {
-  return std::visit(
+  return swl::visit(
       [](auto value) -> FCesiumPropertyArray {
         if constexpr (CesiumGltf::IsMetadataArray<decltype(value)>::value) {
           return FCesiumPropertyArray(value);
@@ -274,7 +274,7 @@ FCesiumPropertyArray UCesiumMetadataValueBlueprintLibrary::GetArray(
 
 bool UCesiumMetadataValueBlueprintLibrary::IsEmpty(
     UPARAM(ref) const FCesiumMetadataValue& Value) {
-  return std::holds_alternative<std::monostate>(Value._value);
+  return swl::holds_alternative<swl::monostate>(Value._value);
 }
 
 TMap<FString, FString> UCesiumMetadataValueBlueprintLibrary::GetValuesAsStrings(