… .github/actions/install-unreal-macos/action.yml
… .github/workflows/build.yml
… .github/workflows/buildAndroid.yml
… .github/workflows/buildApple.yml
… .github/workflows/buildLinux.yml
… .github/workflows/buildMac.yml
… .github/workflows/buildWindows.yml
… Source/CesiumRuntime/Private/CesiumPropertyArrayBlueprintLibrary.cpp
… Source/CesiumRuntime/Private/CesiumPropertyTableProperty.cpp
… Source/CesiumRuntime/Private/CesiumPropertyTextureProperty.cpp
… Source/CesiumRuntime/Private/Tests/CesiumLoadTestCore.h
… Source/CesiumRuntime/Private/Tests/CesiumLoadTestSamples.cpp
… Source/CesiumRuntime/Private/Tests/CesiumMetadataValue.spec.cpp
… Source/CesiumRuntime/Private/Tests/CesiumPropertyArray.spec.cpp
… Source/CesiumRuntime/Private/Tests/Google3dTilesLoadTest.cpp
… Source/CesiumRuntime/Private/UnrealAssetAccessor.cpp
… Source/CesiumRuntime/Public/CesiumMetadataValue.h
… Source/CesiumRuntime/Public/CesiumPropertyArray.h
… extern/CMakeLists.txt
… extern/vcpkg-overlays/triplets/shared/common.cmake