In my limited tests this drastically improves compile times.

So far cesium-native is still using C++17 and std::variant. I've only moved cesium-unreal to C++20 (it was already there for UE5.4, but now 5.2 and 5.3 build with C++20 too). And because [swl::variant](https://github.com/groundswellaudio/swl-variant) requires C++20, I've only moved the variants in the Unreal-specific code to swl::variant. But even with that limited change, Linux compile times on my system were cut by 35%.

This is a PR into #1447 which I'll probably merge myself if CI is happy on all platforms.