@csciguy8 2024-06-14
>The complete plugin build on Linux went from 28 minutes 19 seconds down to 20 minutes 8 seconds.
> The size of the Linux binaries in this branch are also significantly smaller. Total ZIP size is down from 188 MB to 167 MB. Uncompressed size goes down from 903 MB to 633 MB.

Nice.  Worth defining `variant` differently per platform to work around the problems?