@jherico 2024-06-14
Possibly caused by this:  https://stackoverflow.com/a/70997041/85306 and also you might be able to work around it with a define of `SWL_VARIANT_NO_CONSTEXPR_EMPLACE`