@kring 2024-06-16
Looks like it makes a huge difference on Windows, too. It goes from 39 min 17 s down to 12 min 23 s.